Added  org-file for literate programming style
Considering the little documentation and as a help for developing the
enhancements I decided to create an org-file to rewrite the program in a
literary style.

I do not intend to throw away the original code but to use it as the
foundation for an improved version with a better UI and better
automated.

This point is the milestone where I includded the description of how and
why the program will behave as it will, based on the requirements I
discussed witht the people at the LAb (Christian)