Added mock-up GUI
With a little of effort I managed to add a mock-up GUI with buttons.
Right now the GUI is absolutely useless, however as it stand I'm allowed
to produce a successfull termination with the GUI.
The next step is to add 3 functional buttons, one for each action of the
program.