Built the basic functionality for UI
The GUI already has its basic constructs properly defined.
I've added buttons to excecute the commands, although right now they set
the corresponding flag.

I've also created a bare skeleton for the configuration of the UI
parameters including a button to configure the directory where to store
the results. (probable just the parent directory, but the capability is
already there)