Fixed minor issue to ensure that the program runs
Forgot to modify all required parameters to ensure that the program
would run after the inheritance. Now the program runs.

Also fixed the print function as it didn't work as expected.

Overloaded the print function for the piezo as we want to print it as an
integer not as a float, but that is almost inconsecuential