Wrote Adwin IV program + improved basic-mode
Wrote and documented the basic ADWIN program to measure the I/V
characteristic of the sample.
This program sets the voltage at the juncture and measures the current
across it.
The program performs continous measurements and is the task of the
external program to extract the relevant information.

Also improved basic mode. The new version correctly highlights the
comments and some keywords including the ADWIN specific variables.