Advanced prototype of the adwin interfaces
I improved the interfaces, I still need to properly define de the
configuration parameters for the instrument measurement process with
that respective interface, and the start/end process features, but the
skeleton is there.

I also moved the "comments" header org argument in order to have them in
more central places and disable them more easily if necessary.