Fixing imports to avoid recursive imports
Changed the ADWIN_GCONST from adwin driver to parameters to avoid
recursive imports and have a more "linear" dependency tree. This
should improve the robustness of the code across python versions.

I also made a small fix to comply with python standars by moving the
imports of system libraries to the top of the import chain.