Testing Save and plot data and fixed
I worked on the save and plot data from the histogram, and managed to
get a working version using the hardware.
Now I'm able to properly plot and save the data in a similar manner as
is done by the old program.

At this point I should be able to remove the old program from the repository

- I have also decided to include the generated files