Added a simple interface for analysis
There are cases where is not possible to break the juncture using the
motor either for a defect on the device or another reason.
In such scenarios manual intervention is necessary, and the user need
a continous measurement of the conductance in order to diagnose the
problem. I have provided with the base interface in order to perform
such measurements.

I also removed some dead code using vulture