Revision of testfiles and histogram process
Double checked some test files and the required cofigurations of the
program.
As it is, it should work properly save for the distance/voltage
calibration.

I Also modified the histogram measurement program as to skip less
datapoints, instead of avoiding skipping them at all because the
over-representation was going a little out of hand.